requirements.txt
streamlit
qrcode
Pillow
google-generativeai